Refactor classes for avoid nests
