+++ paper-api/src/main/java/io/papermc/paper/datacomponent/item/BlocksAttacks.java
@@ -1,13 +1,11 @@
 package io.papermc.paper.datacomponent.item;
 
 import io.papermc.paper.datacomponent.DataComponentBuilder;
-import io.papermc.paper.registry.set.RegistryKeySet;
+import io.papermc.paper.datacomponent.item.blocksattacks.DamageReduction;
+import io.papermc.paper.datacomponent.item.blocksattacks.ItemDamageFunction;
 import io.papermc.paper.registry.tag.TagKey;
 import net.kyori.adventure.key.Key;
 import org.bukkit.damage.DamageType;
-import org.checkerframework.checker.index.qual.NonNegative;
-import org.checkerframework.checker.index.qual.Positive;
-import org.checkerframework.common.value.qual.IntRange;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Contract;
 import org.jspecify.annotations.NullMarked;
@@ -71,82 +69,5 @@ public interface BlocksAttacks {
 
         @Contract(value = "_ -> this", mutates = "this")
         Builder disableSound(@Nullable Key sound);
-    }
-
-    @ApiStatus.Experimental
-    @ApiStatus.NonExtendable
-    interface DamageReduction {
-
-        @Contract(value = "-> new", pure = true)
-        static DamageReduction.Builder damageReduction() {
-            return ItemComponentTypesBridge.bridge().blocksAttacksDamageReduction();
-        }
-
-        @Nullable
-        RegistryKeySet<DamageType> type();
-
-        @Positive
-        float horizontalBlockingAngle();
-
-        float base();
-
-        float factor();
-
-        /**
-         * Builder for {@link BlocksAttacks.DamageReduction}.
-         */
-        @ApiStatus.Experimental
-        @ApiStatus.NonExtendable
-        interface Builder extends DataComponentBuilder<DamageReduction> {
-
-            @Contract(value = "_ -> this", mutates = "this")
-            BlocksAttacks.DamageReduction.Builder type(RegistryKeySet<DamageType> type);
-
-            @Contract(value = "_ -> this", mutates = "this")
-            BlocksAttacks.DamageReduction.Builder horizontalBlockingAngle(@Positive float horizontalBlockingAngle);
-
-            @Contract(value = "_ -> this", mutates = "this")
-            BlocksAttacks.DamageReduction.Builder base(float base);
-
-            @Contract(value = "_ -> this", mutates = "this")
-            BlocksAttacks.DamageReduction.Builder factor(float factor);
-        }
-
-    }
-
-    @ApiStatus.Experimental
-    @ApiStatus.NonExtendable
-    interface ItemDamageFunction {
-
-        @Contract(value = "-> new", pure = true)
-        static ItemDamageFunction.Builder itemDamageFunction() {
-            return ItemComponentTypesBridge.bridge().blocksAttacksItemDamageFunction();
-        }
-
-        @NonNegative
-        float threshold();
-
-        float base();
-
-        float factor();
-
-        int damageToApply(float damage);
-
-        /**
-         * Builder for {@link BlocksAttacks.DamageReduction}.
-         */
-        @ApiStatus.Experimental
-        @ApiStatus.NonExtendable
-        interface Builder extends DataComponentBuilder<ItemDamageFunction> {
-
-            @Contract(value = "_ -> this", mutates = "this")
-            BlocksAttacks.ItemDamageFunction.Builder threshold(@NonNegative final float threshold);
-
-            @Contract(value = "_ -> this", mutates = "this")
-            BlocksAttacks.ItemDamageFunction.Builder base(final float base);
-
-            @Contract(value = "_ -> this", mutates = "this")
-            BlocksAttacks.ItemDamageFunction.Builder factor(final float factor);
-        }
     }
 }
