+++ paper-server/src/main/java/io/papermc/paper/datacomponent/item/ItemComponentTypesBridgesImpl.java
@@ -238,16 +238,6 @@ public final class ItemComponentTypesBridgesImpl implements ItemComponentTypesBr
     }
 
     @Override
-    public BlocksAttacks.DamageReduction.Builder blocksAttacksDamageReduction() {
-        return new PaperBlocksAttacks.PaperDamageReduction.BuilderImpl();
-    }
-
-    @Override
-    public BlocksAttacks.ItemDamageFunction.Builder blocksAttacksItemDamageFunction() {
-        return new PaperBlocksAttacks.PaperItemDamageFunction.BuilderImpl();
-    }
-
-    @Override
     public TooltipDisplay.Builder tooltipDisplay() {
         return new PaperTooltipDisplay.BuilderImpl();
     }
