+++ paper-server/src/main/java/io/papermc/paper/datacomponent/item/PaperBlocksAttacks.java
@@ -2,21 +2,18 @@ package io.papermc.paper.datacomponent.item;
 
 import com.google.common.base.Preconditions;
 import io.papermc.paper.adventure.PaperAdventure;
+import io.papermc.paper.datacomponent.item.blocksattacks.DamageReduction;
+import io.papermc.paper.datacomponent.item.blocksattacks.ItemDamageFunction;
+import io.papermc.paper.datacomponent.item.blocksattacks.PaperDamageReduction;
+import io.papermc.paper.datacomponent.item.blocksattacks.PaperItemDamageFunction;
 import io.papermc.paper.registry.PaperRegistries;
-import io.papermc.paper.registry.RegistryKey;
-import io.papermc.paper.registry.set.PaperRegistrySets;
-import io.papermc.paper.registry.set.RegistryKeySet;
 import io.papermc.paper.registry.tag.TagKey;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import net.kyori.adventure.key.Key;
-import net.minecraft.core.HolderSet;
-import net.minecraft.core.registries.Registries;
 import org.bukkit.craftbukkit.util.Handleable;
 import org.bukkit.damage.DamageType;
-import org.checkerframework.checker.index.qual.NonNegative;
-import org.checkerframework.checker.index.qual.Positive;
 import org.jetbrains.annotations.Nullable;
 
 public record PaperBlocksAttacks(
@@ -136,145 +133,6 @@ public record PaperBlocksAttacks(
                 Optional.ofNullable(this.blockSound).map(PaperAdventure::resolveSound),
                 Optional.ofNullable(this.disableSound).map(PaperAdventure::resolveSound)
             ));
-        }
-    }
-
-    public record PaperDamageReduction(
-        net.minecraft.world.item.component.BlocksAttacks.DamageReduction impl
-    ) implements DamageReduction, Handleable<net.minecraft.world.item.component.BlocksAttacks.DamageReduction> {
-
-        @Override
-        public net.minecraft.world.item.component.BlocksAttacks.DamageReduction getHandle() {
-            return this.impl;
-        }
-
-        @Override
-        public @Nullable RegistryKeySet<DamageType> type() {
-            return this.impl.type().map((set) -> PaperRegistrySets.convertToApi(RegistryKey.DAMAGE_TYPE, set)).orElse(null);
-        }
-
-        @Override
-        public @Positive float horizontalBlockingAngle() {
-            return this.impl.horizontalBlockingAngle();
-        }
-
-        @Override
-        public float base() {
-            return this.impl.base();
-        }
-
-        @Override
-        public float factor() {
-            return this.impl.factor();
-        }
-
-        static final class BuilderImpl implements Builder {
-
-            private Optional<HolderSet<net.minecraft.world.damagesource.DamageType>> type = Optional.empty();
-            private float horizontalBlockingAngle = 90f;
-            private float base = 0;
-            private float factor = 0;
-
-            @Override
-            public Builder type(final @Nullable RegistryKeySet<DamageType> type) {
-                this.type = Optional.ofNullable(type)
-                    .map((set) -> PaperRegistrySets.convertToNms(Registries.DAMAGE_TYPE, net.minecraft.server.MinecraftServer.getServer().registryAccess().createSerializationContext(net.minecraft.nbt.NbtOps.INSTANCE).lookupProvider, set));
-                return this;
-            }
-
-            @Override
-            public Builder horizontalBlockingAngle(@Positive final float horizontalBlockingAngle) {
-                Preconditions.checkArgument(horizontalBlockingAngle > 0, "horizontalBlockingAngle must be positive and not zero, was %s", horizontalBlockingAngle);
-                this.horizontalBlockingAngle = horizontalBlockingAngle;
-                return this;
-            }
-
-            @Override
-            public Builder base(final float base) {
-                this.base = base;
-                return this;
-            }
-
-            @Override
-            public Builder factor(final float factor) {
-                this.factor = factor;
-                return this;
-            }
-
-            @Override
-            public DamageReduction build() {
-                return new PaperDamageReduction(new net.minecraft.world.item.component.BlocksAttacks.DamageReduction(
-                    this.horizontalBlockingAngle,
-                    this.type,
-                    this.base,
-                    this.factor
-                ));
-            }
-        }
-    }
-
-    public record PaperItemDamageFunction(
-        net.minecraft.world.item.component.BlocksAttacks.ItemDamageFunction impl
-    ) implements ItemDamageFunction, Handleable<net.minecraft.world.item.component.BlocksAttacks.ItemDamageFunction> {
-
-        @Override
-        public net.minecraft.world.item.component.BlocksAttacks.ItemDamageFunction getHandle() {
-            return this.impl;
-        }
-
-        @Override
-        public @NonNegative float threshold() {
-            return this.impl.threshold();
-        }
-
-        @Override
-        public float base() {
-            return this.impl.base();
-        }
-
-        @Override
-        public float factor() {
-            return this.impl.factor();
-        }
-
-        @Override
-        public int damageToApply(final float damage) {
-            return this.impl.apply(damage);
-        }
-
-        static final class BuilderImpl implements Builder {
-
-            private float threshold;
-            private float base;
-            private float factor;
-
-            @Override
-            public Builder threshold(@NonNegative final float threshold) {
-                Preconditions.checkArgument(threshold >= 0, "threshold must be non-negative, was %s", threshold);
-                this.threshold = threshold;
-                return this;
-            }
-
-            @Override
-            public Builder base(final float base) {
-                this.base = base;
-                return this;
-            }
-
-            @Override
-            public Builder factor(final float factor) {
-                this.factor = factor;
-                return this;
-            }
-
-            @Override
-            public ItemDamageFunction build() {
-                return new PaperItemDamageFunction(new net.minecraft.world.item.component.BlocksAttacks.ItemDamageFunction(
-                    this.threshold,
-                    this.base,
-                    this.factor
-                ));
-            }
         }
     }
 }
