+++ paper-server/src/main/java/io/papermc/paper/datacomponent/item/blocksattacks/BlocksAttacksBridgesImpl.java
@@ -1,0 +1,14 @@
+package io.papermc.paper.datacomponent.item.blocksattacks;
+
+public final class BlocksAttacksBridgesImpl implements BlocksAttacksBridge {
+
+    @Override
+    public DamageReduction.Builder blocksAttacksDamageReduction() {
+        return new PaperDamageReduction.BuilderImpl();
+    }
+
+    @Override
+    public ItemDamageFunction.Builder blocksAttacksItemDamageFunction() {
+        return new PaperItemDamageFunction.BuilderImpl();
+    }
+}
