+++ paper-server/src/main/java/io/papermc/paper/datacomponent/item/blocksattacks/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Relating to block attacks for components.
+ */
+@NullMarked
+package io.papermc.paper.datacomponent.item.blocksattacks;
+
+import org.jspecify.annotations.NullMarked;
