+++ paper-server/src/main/resources/META-INF/services/io.papermc.paper.datacomponent.item.blocksattacks.BlocksAttacksBridge
@@ -1,0 +1,1 @@
+io.papermc.paper.datacomponent.item.blocksattacks.BlocksAttacksBridge
